Rename LocaleType til Locale
