+++ src/common/typer/kontrakt/generelle.ts
@@ -1,7 +1,7 @@
 import { Alpha3Code } from 'i18n-iso-countries';
 
 import { LocaleRecordBlock, LocaleRecordString } from '../../../frontend/typer/common';
-import { LocaleType } from '../localeType';
+import { LocaleType } from '../locale';
 
 export interface ISøknadsfelt<T> {
     label: Record<LocaleType, string>;
