+++ src/common/typer/locale.ts
