+++ src/frontend/utils/dato.tsx
@@ -21,7 +21,7 @@ import { feil, type FeltState, ok } from '@navikt/familie-skjema';
 import { AlternativtSvarForInput, DatoMedUkjent, LocaleRecordBlock } from '../typer/common';
 import { PlainTekst } from '../../common/typer/kontrakt/generelle';
 import { IFormateringsfeilmeldingerTekstinnhold } from '../typer/sanity/tekstInnhold';
-import { LocaleType } from '../../common/typer/localeType';
+import { LocaleType } from '../../common/typer/locale';
 import { ISODateString } from '../../common/typer/ISODateString';
 
 export const erDatoFormatGodkjent = (dato: Date) => isValid(dato);
