+++ src/frontend/utils/mappingTilKontrakt/hjelpefunksjoner.ts
@@ -2,7 +2,7 @@ import { ESvar } from '@navikt/familie-form-elements';
 
 import { AlternativtSvarForInput, LocaleRecordBlock, LocaleRecordString } from '../../typer/common';
 import { FlettefeltVerdier, ISøknadsfelt, TilRestLocaleRecord } from '../../../common/typer/kontrakt/generelle';
-import { LocaleType } from '../../../common/typer/localeType';
+import { LocaleType } from '../../../common/typer/locale';
 
 export const søknadsfeltHof =
     (tilRestLocaleRecord: TilRestLocaleRecord) =>
